+++ cpp/include/graph.hpp
@@ -14,6 +14,10 @@
  * limitations under the License.
  */
 #pragma once
+#include <comms_mpi.hpp>
+#include <iostream>
+#include <memory>
+#include <rmm/device_buffer.hpp>
 
 namespace cugraph {
 namespace experimental {
@@ -45,8 +49,9 @@ enum class DegreeDirection {
  * @tparam WT   Type of weight
  */
 template <typename VT, typename ET, typename WT>
-class GraphBase {
+class GraphViewBase {
  public:
+  Comm comm;
   WT *edge_data;  ///< edge weight
 
   GraphProperties prop;
@@ -57,17 +62,21 @@ class GraphBase {
   /**
    * @brief      Fill the identifiers array with the vertex identifiers.
    *
-   * @param[out]    identifier      Pointer to device memory to store the vertex identifiers
+   * @param[out]    identifier      Pointer to device memory to store the vertex
+   * identifiers
    */
   void get_vertex_identifiers(VT *identifiers) const;
+  void set_communicator(Comm &comm_) { comm = comm_; }
 
-  GraphBase(WT *edge_data_, VT number_of_vertices_, ET number_of_edges_)
+  GraphViewBase(WT *edge_data_, VT number_of_vertices_, ET number_of_edges_)
     : edge_data(edge_data_),
+      comm(),
       prop(),
       number_of_vertices(number_of_vertices_),
       number_of_edges(number_of_edges_)
   {
   }
+  bool has_data(void) const { return edge_data != nullptr; }
 };
 
 /**
@@ -78,7 +87,7 @@ class GraphBase {
  * @tparam WT   Type of weight
  */
 template <typename VT, typename ET, typename WT>
-class GraphCOO : public GraphBase<VT, ET, WT> {
+class GraphCOOView : public GraphViewBase<VT, ET, WT> {
  public:
   VT *src_indices{nullptr};  ///< rowInd
   VT *dst_indices{nullptr};  ///< colInd
@@ -97,12 +106,12 @@ class GraphCOO : public GraphBase<VT, ET, WT> {
   /**
    * @brief      Default constructor
    */
-  GraphCOO() : GraphBase<VT, ET, WT>(nullptr, 0, 0) {}
+  GraphCOOView() : GraphViewBase<VT, ET, WT>(nullptr, 0, 0) {}
 
   /**
    * @brief      Wrap existing arrays representing an edge list in a Graph.
    *
-   *             GraphCOO does not own the memory used to represent this graph. This
+   *             GraphCOOView does not own the memory used to represent this graph. This
    *             function does not allocate memory.
    *
    * @param  source_indices        This array of size E (number of edges) contains the index of the
@@ -114,9 +123,9 @@ class GraphCOO : public GraphBase<VT, ET, WT> {
    * @param  number_of_vertices    The number of vertices in the graph
    * @param  number_of_edges       The number of edges in the graph
    */
-  GraphCOO(
+  GraphCOOView(
     VT *src_indices_, VT *dst_indices_, WT *edge_data_, VT number_of_vertices_, ET number_of_edges_)
-    : GraphBase<VT, ET, WT>(edge_data_, number_of_vertices_, number_of_edges_),
+    : GraphViewBase<VT, ET, WT>(edge_data_, number_of_vertices_, number_of_edges_),
       src_indices(src_indices_),
       dst_indices(dst_indices_)
   {
@@ -132,15 +141,17 @@ class GraphCOO : public GraphBase<VT, ET, WT> {
  * @tparam WT   Type of weight
  */
 template <typename VT, typename ET, typename WT>
-class GraphCompressedSparseBase : public GraphBase<VT, ET, WT> {
+class GraphCompressedSparseBaseView : public GraphViewBase<VT, ET, WT> {
  public:
   ET *offsets{nullptr};  ///< CSR offsets
   VT *indices{nullptr};  ///< CSR indices
 
   /**
-   * @brief      Fill the identifiers in the array with the source vertex identifiers
+   * @brief      Fill the identifiers in the array with the source vertex
+   * identifiers
    *
-   * @param[out]    src_indices      Pointer to device memory to store the source vertex identifiers
+   * @param[out]    src_indices      Pointer to device memory to store the
+   * source vertex identifiers
    */
   void get_source_indices(VT *src_indices) const;
 
@@ -160,7 +171,7 @@ class GraphCompressedSparseBase : public GraphBase<VT, ET, WT> {
 
   /**
    * @brief      Wrap existing arrays representing adjacency lists in a Graph.
-   *             GraphCSR does not own the memory used to represent this graph. This
+   *             GraphCSRView does not own the memory used to represent this graph. This
    *             function does not allocate memory.
    *
    * @param  offsets               This array of size V+1 (V is number of vertices) contains the
@@ -173,9 +184,9 @@ class GraphCompressedSparseBase : public GraphBase<VT, ET, WT> {
    * @param  number_of_vertices    The number of vertices in the graph
    * @param  number_of_edges       The number of edges in the graph
    */
-  GraphCompressedSparseBase(
+  GraphCompressedSparseBaseView(
     ET *offsets_, VT *indices_, WT *edge_data_, VT number_of_vertices_, ET number_of_edges_)
-    : GraphBase<VT, ET, WT>(edge_data_, number_of_vertices_, number_of_edges_),
+    : GraphViewBase<VT, ET, WT>(edge_data_, number_of_vertices_, number_of_edges_),
       offsets{offsets_},
       indices{indices_}
   {
@@ -190,16 +201,16 @@ class GraphCompressedSparseBase : public GraphBase<VT, ET, WT> {
  * @tparam WT   Type of weight
  */
 template <typename VT, typename ET, typename WT>
-class GraphCSR : public GraphCompressedSparseBase<VT, ET, WT> {
+class GraphCSRView : public GraphCompressedSparseBaseView<VT, ET, WT> {
  public:
   /**
    * @brief      Default constructor
    */
-  GraphCSR() : GraphCompressedSparseBase<VT, ET, WT>(nullptr, nullptr, nullptr, 0, 0) {}
+  GraphCSRView() : GraphCompressedSparseBaseView<VT, ET, WT>(nullptr, nullptr, nullptr, 0, 0) {}
 
   /**
    * @brief      Wrap existing arrays representing adjacency lists in a Graph.
-   *             GraphCSR does not own the memory used to represent this graph. This
+   *             GraphCSRView does not own the memory used to represent this graph. This
    *             function does not allocate memory.
    *
    * @param  offsets               This array of size V+1 (V is number of vertices) contains the
@@ -212,8 +223,9 @@ class GraphCSR : public GraphCompressedSparseBase<VT, ET, WT> {
    * @param  number_of_vertices    The number of vertices in the graph
    * @param  number_of_edges       The number of edges in the graph
    */
-  GraphCSR(ET *offsets_, VT *indices_, WT *edge_data_, VT number_of_vertices_, ET number_of_edges_)
-    : GraphCompressedSparseBase<VT, ET, WT>(
+  GraphCSRView(
+    ET *offsets_, VT *indices_, WT *edge_data_, VT number_of_vertices_, ET number_of_edges_)
+    : GraphCompressedSparseBaseView<VT, ET, WT>(
         offsets_, indices_, edge_data_, number_of_vertices_, number_of_edges_)
   {
   }
@@ -227,16 +239,16 @@ class GraphCSR : public GraphCompressedSparseBase<VT, ET, WT> {
  * @tparam WT   Type of weight
  */
 template <typename VT, typename ET, typename WT>
-class GraphCSC : public GraphCompressedSparseBase<VT, ET, WT> {
+class GraphCSCView : public GraphCompressedSparseBaseView<VT, ET, WT> {
  public:
   /**
    * @brief      Default constructor
    */
-  GraphCSC() : GraphCompressedSparseBase<VT, ET, WT>(nullptr, nullptr, nullptr, 0, 0) {}
+  GraphCSCView() : GraphCompressedSparseBaseView<VT, ET, WT>(nullptr, nullptr, nullptr, 0, 0) {}
 
   /**
    * @brief      Wrap existing arrays representing transposed adjacency lists in a Graph.
-   *             GraphCSC does not own the memory used to represent this graph. This
+   *             GraphCSCView does not own the memory used to represent this graph. This
    *             function does not allocate memory.
    *
    * @param  offsets               This array of size V+1 (V is number of vertices) contains the
@@ -249,10 +261,310 @@ class GraphCSC : public GraphCompressedSparseBase<VT, ET, WT> {
    * @param  number_of_vertices    The number of vertices in the graph
    * @param  number_of_edges       The number of edges in the graph
    */
-  GraphCSC(ET *offsets_, VT *indices_, WT *edge_data_, VT number_of_vertices_, ET number_of_edges_)
-    : GraphCompressedSparseBase<VT, ET, WT>(
+  GraphCSCView(
+    ET *offsets_, VT *indices_, WT *edge_data_, VT number_of_vertices_, ET number_of_edges_)
+    : GraphCompressedSparseBaseView<VT, ET, WT>(
         offsets_, indices_, edge_data_, number_of_vertices_, number_of_edges_)
   {
+  }
+};
+
+/**
+ * @brief      TODO : Change this Take ownership of the provided graph arrays in COO format
+ *
+ * @param  source_indices        This array of size E (number of edges) contains the index of the
+ * source for each edge. Indices must be in the range [0, V-1].
+ * @param  destination_indices   This array of size E (number of edges) contains the index of the
+ * destination for each edge. Indices must be in the range [0, V-1].
+ * @param  edge_data             This array size E (number of edges) contains the weight for each
+ * edge.  This array can be null in which case the graph is considered unweighted.
+ * @param  number_of_vertices    The number of vertices in the graph
+ * @param  number_of_edges       The number of edges in the graph
+ */
+template <typename VT, typename ET, typename WT>
+struct GraphCOOContents {
+  VT number_of_vertices;
+  ET number_of_edges;
+  std::unique_ptr<rmm::device_buffer> src_indices;
+  std::unique_ptr<rmm::device_buffer> dst_indices;
+  std::unique_ptr<rmm::device_buffer> edge_data;
+};
+
+/**
+ * @brief       A constructed graph stored in COO (COOrdinate) format.
+ *
+ * This class will src_indices and dst_indicies (until moved)
+ *
+ * @tparam VT   Type of vertex id
+ * @tparam ET   Type of edge id
+ * @tparam WT   Type of weight
+ */
+template <typename VT, typename ET, typename WT>
+class GraphCOO {
+  VT number_of_vertices_;
+  ET number_of_edges_;
+  rmm::device_buffer src_indices_{};  ///< rowInd
+  rmm::device_buffer dst_indices_{};  ///< colInd
+  rmm::device_buffer edge_data_{};    ///< CSR data
+
+ public:
+  /**
+   * @brief      Take ownership of the provided graph arrays in COO format
+   *
+   * @param  source_indices        This array of size E (number of edges) contains the index of the
+   * source for each edge. Indices must be in the range [0, V-1].
+   * @param  destination_indices   This array of size E (number of edges) contains the index of the
+   * destination for each edge. Indices must be in the range [0, V-1].
+   * @param  edge_data             This array size E (number of edges) contains the weight for each
+   * edge.  This array can be null in which case the graph is considered unweighted.
+   * @param  number_of_vertices    The number of vertices in the graph
+   * @param  number_of_edges       The number of edges in the graph
+   */
+  GraphCOO(VT number_of_vertices,
+           ET number_of_edges,
+           bool has_data                       = false,
+           cudaStream_t stream                 = nullptr,
+           rmm::mr::device_memory_resource *mr = rmm::mr::get_default_resource())
+    : number_of_vertices_(number_of_vertices),
+      number_of_edges_(number_of_edges),
+      src_indices_(sizeof(VT) * number_of_edges, stream, mr),
+      dst_indices_(sizeof(VT) * number_of_edges, stream, mr),
+      edge_data_((has_data ? sizeof(WT) * number_of_edges : 0), stream, mr)
+  {
+  }
+
+  GraphCOO(GraphCOOView<VT, ET, WT> const &graph,
+           cudaStream_t stream                 = nullptr,
+           rmm::mr::device_memory_resource *mr = rmm::mr::get_default_resource())
+    : number_of_vertices_(graph.number_of_vertices),
+      number_of_edges_(graph.number_of_edges),
+      src_indices_(graph.src_indices, graph.number_of_edges * sizeof(VT), stream, mr),
+      dst_indices_(graph.dst_indices, graph.number_of_edges * sizeof(VT), stream, mr)
+  {
+    if (graph.has_data()) {
+      edge_data_ =
+        rmm::device_buffer{graph.edge_data, graph.number_of_edges * sizeof(WT), stream, mr};
+    }
+  }
+
+  VT number_of_vertices(void) { return number_of_vertices_; }
+  ET number_of_edges(void) { return number_of_edges_; }
+  VT *src_indices(void) { return static_cast<VT *>(src_indices_.data()); }
+  VT *dst_indices(void) { return static_cast<VT *>(dst_indices_.data()); }
+  WT *edge_data(void) { return static_cast<WT *>(edge_data_.data()); }
+
+  GraphCOOContents<VT, ET, WT> release() noexcept
+  {
+    VT number_of_vertices = number_of_vertices_;
+    ET number_of_edges    = number_of_edges_;
+    number_of_vertices_   = 0;
+    number_of_edges_      = 0;
+    return GraphCOOContents<VT, ET, WT>{
+      number_of_vertices,
+      number_of_edges,
+      std::make_unique<rmm::device_buffer>(std::move(src_indices_)),
+      std::make_unique<rmm::device_buffer>(std::move(dst_indices_)),
+      std::make_unique<rmm::device_buffer>(std::move(edge_data_))};
+  }
+
+  GraphCOOView<VT, ET, WT> view(void) noexcept
+  {
+    return GraphCOOView<VT, ET, WT>(
+      src_indices(), dst_indices(), edge_data(), number_of_vertices_, number_of_edges_);
+  }
+
+  bool has_data(void) { return nullptr != edge_data_.data(); }
+};
+
+template <typename VT, typename ET, typename WT>
+struct GraphSparseContents {
+  VT number_of_vertices;
+  ET number_of_edges;
+  std::unique_ptr<rmm::device_buffer> offsets;
+  std::unique_ptr<rmm::device_buffer> indices;
+  std::unique_ptr<rmm::device_buffer> edge_data;
+};
+
+/**
+ * @brief       Base class for constructted graphs stored in CSR (Compressed Sparse Row) format or
+ * CSC (Compressed Sparse Column) format
+ *
+ * @tparam VT   Type of vertex id
+ * @tparam ET   Type of edge id
+ * @tparam WT   Type of weight
+ */
+template <typename VT, typename ET, typename WT>
+class GraphCompressedSparseBase {
+  VT number_of_vertices_{0};
+  ET number_of_edges_{0};
+  rmm::device_buffer offsets_{};    ///< CSR offsets
+  rmm::device_buffer indices_{};    ///< CSR indices
+  rmm::device_buffer edge_data_{};  ///< CSR data
+
+  bool has_data_{false};
+
+ public:
+  /**
+   * @brief      Take ownership of the provided graph arrays in CSR/CSC format
+   *
+   * @param  offsets               This array of size V+1 (V is number of vertices) contains the
+   * offset of adjacency lists of every vertex. Offsets must be in the range [0, E] (number of
+   * edges).
+   * @param  indices               This array of size E contains the index of the destination for
+   * each edge. Indices must be in the range [0, V-1].
+   * @param  edge_data             This array of size E (number of edges) contains the weight for
+   * each edge.  This array can be null in which case the graph is considered unweighted.
+   * @param  number_of_vertices    The number of vertices in the graph
+   * @param  number_of_edges       The number of edges in the graph
+   */
+  GraphCompressedSparseBase(VT number_of_vertices,
+                            ET number_of_edges,
+                            bool has_data,
+                            cudaStream_t stream,
+                            rmm::mr::device_memory_resource *mr)
+    : number_of_vertices_(number_of_vertices),
+      number_of_edges_(number_of_edges),
+      offsets_(sizeof(ET) * (number_of_vertices + 1), stream, mr),
+      indices_(sizeof(VT) * number_of_edges, stream, mr),
+      edge_data_((has_data ? sizeof(WT) * number_of_edges : 0), stream, mr)
+  {
+  }
+
+  GraphCompressedSparseBase(GraphSparseContents<VT, ET, WT> &&contents)
+    : number_of_vertices_(contents.number_of_vertices),
+      number_of_edges_(contents.number_of_edges),
+      offsets_(std::move(*contents.offsets.release())),
+      indices_(std::move(*contents.indices.release())),
+      edge_data_(std::move(*contents.edge_data.release()))
+  {
+  }
+
+  VT number_of_vertices(void) { return number_of_vertices_; }
+  ET number_of_edges(void) { return number_of_edges_; }
+  ET *offsets(void) { return static_cast<ET *>(offsets_.data()); }
+  VT *indices(void) { return static_cast<VT *>(indices_.data()); }
+  WT *edge_data(void) { return static_cast<WT *>(edge_data_.data()); }
+
+  GraphSparseContents<VT, ET, WT> release() noexcept
+  {
+    VT number_of_vertices = number_of_vertices_;
+    ET number_of_edges    = number_of_edges_;
+    number_of_vertices_   = 0;
+    number_of_edges_      = 0;
+    return GraphSparseContents<VT, ET, WT>{
+      number_of_vertices,
+      number_of_edges,
+      std::make_unique<rmm::device_buffer>(std::move(offsets_)),
+      std::make_unique<rmm::device_buffer>(std::move(indices_)),
+      std::make_unique<rmm::device_buffer>(std::move(edge_data_))};
+  }
+
+  bool has_data(void) { return nullptr != edge_data_.data(); }
+};
+
+/**
+ * @brief       A constructed graph stored in CSR (Compressed Sparse Row) format.
+ *
+ * @tparam VT   Type of vertex id
+ * @tparam ET   Type of edge id
+ * @tparam WT   Type of weight
+ */
+template <typename VT, typename ET, typename WT>
+class GraphCSR : public GraphCompressedSparseBase<VT, ET, WT> {
+ public:
+  /**
+   * @brief      Default constructor
+   */
+  GraphCSR() : GraphCompressedSparseBase<VT, ET, WT>() {}
+
+  /**
+   * @brief      Take ownership of the provided graph arrays in CSR format
+   *
+   * @param  offsets               This array of size V+1 (V is number of vertices) contains the
+   * offset of adjacency lists of every vertex. Offsets must be in the range [0, E] (number of
+   * edges).
+   * @param  indices               This array of size E contains the index of the destination for
+   * each edge. Indices must be in the range [0, V-1].
+   * @param  edge_data             This array of size E (number of edges) contains the weight for
+   * each edge.  This array can be null in which case the graph is considered unweighted.
+   * @param  number_of_vertices    The number of vertices in the graph
+   * @param  number_of_edges       The number of edges in the graph
+   */
+  GraphCSR(VT number_of_vertices_,
+           ET number_of_edges_,
+           bool has_data_                      = false,
+           cudaStream_t stream                 = nullptr,
+           rmm::mr::device_memory_resource *mr = rmm::mr::get_default_resource())
+    : GraphCompressedSparseBase<VT, ET, WT>(
+        number_of_vertices_, number_of_edges_, has_data_, stream, mr)
+  {
+  }
+
+  GraphCSR(GraphSparseContents<VT, ET, WT> &&contents)
+    : GraphCompressedSparseBase<VT, ET, WT>(std::move(contents))
+  {
+  }
+
+  GraphCSRView<VT, ET, WT> view(void) noexcept
+  {
+    return GraphCSRView<VT, ET, WT>(GraphCompressedSparseBase<VT, ET, WT>::offsets(),
+                                    GraphCompressedSparseBase<VT, ET, WT>::indices(),
+                                    GraphCompressedSparseBase<VT, ET, WT>::edge_data(),
+                                    GraphCompressedSparseBase<VT, ET, WT>::number_of_vertices(),
+                                    GraphCompressedSparseBase<VT, ET, WT>::number_of_edges());
+  }
+};
+
+/**
+ * @brief       A constructed graph stored in CSC (Compressed Sparse Column) format.
+ *
+ * @tparam VT   Type of vertex id
+ * @tparam ET   Type of edge id
+ * @tparam WT   Type of weight
+ */
+template <typename VT, typename ET, typename WT>
+class GraphCSC : public GraphCompressedSparseBase<VT, ET, WT> {
+ public:
+  /**
+   * @brief      Default constructor
+   */
+  GraphCSC() : GraphCompressedSparseBase<VT, ET, WT>() {}
+
+  /**
+   * @brief      Take ownership of the provided graph arrays in CSR format
+   *
+   * @param  offsets               This array of size V+1 (V is number of vertices) contains the
+   * offset of adjacency lists of every vertex. Offsets must be in the range [0, E] (number of
+   * edges).
+   * @param  indices               This array of size E contains the index of the destination for
+   * each edge. Indices must be in the range [0, V-1].
+   * @param  edge_data             This array of size E (number of edges) contains the weight for
+   * each edge.  This array can be null in which case the graph is considered unweighted.
+   * @param  number_of_vertices    The number of vertices in the graph
+   * @param  number_of_edges       The number of edges in the graph
+   */
+  GraphCSC(VT number_of_vertices_,
+           ET number_of_edges_,
+           bool has_data_                      = false,
+           cudaStream_t stream                 = nullptr,
+           rmm::mr::device_memory_resource *mr = rmm::mr::get_default_resource())
+    : GraphCompressedSparseBase<VT, ET, WT>(
+        number_of_vertices_, number_of_edges_, has_data_, stream, mr)
+  {
+  }
+
+  GraphCSC(GraphSparseContents<VT, ET, WT> &&contents)
+    : GraphCompressedSparseBase<VT, ET, WT>(contents)
+  {
+  }
+
+  GraphCSCView<VT, ET, WT> view(void) noexcept
+  {
+    return GraphCSCView<VT, ET, WT>(GraphCompressedSparseBase<VT, ET, WT>::offsets(),
+                                    GraphCompressedSparseBase<VT, ET, WT>::indices(),
+                                    GraphCompressedSparseBase<VT, ET, WT>::edge_data(),
+                                    GraphCompressedSparseBase<VT, ET, WT>::number_of_vertices(),
+                                    GraphCompressedSparseBase<VT, ET, WT>::number_of_edges());
   }
 };